快速启动脚本修改，不再依赖本地zk
