+++ bistoury-proxy/conf/releaseInfo_config.properties
@@ -2,7 +2,7 @@
 #发布信息，应用的发布信息文件的路径，默认值为相对日志目录的相对路径../webapps/releaseInfo.properties，可以通过default修改默认值
 #可以是绝对路径
 #可选，配置默认路径，默认值为 ../webapps/releaseInfo.properties
-default=/tmp/releaseInfo.properties
+default=/tmp/bistoury/releaseInfo.properties
 #也可以是相对日志目录的相对路径
 #default=../webapps/releaseInfo.properties
 #针对appcode进行设置，即每个appcode的发布信息文件路径需要保持一致，如：
