+++ bistoury-proxy/conf/server.properties
@@ -1,3 +1,3 @@
 tomcat.port=8080
-tomcat.basedir=/tmp/tomcat
+tomcat.basedir=/tmp/bistoury/tomcat/proxy
