+++ bistoury-ui/conf/server.properties
@@ -1,3 +1,3 @@
 tomcat.port=8081
-tomcat.basedir=/tmp/tomcat
+tomcat.basedir=/tmp/bistoury/tomcat/ui
