+++ bistoury-ui/src/bin/bistoury-ui-env.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
-JAVA_HOME="/tmp/java"
+JAVA_HOME="/tmp/bistoury/java"
 JAVA_OPTS="-Dbistoury.conf=$BISTOURY_COF_DIR"
